requirements.txt
numpy
rasterio
matplotlib
torch
torchvision
timm
rasterio
pyyaml
